requirements.txt
aiogram>=3.0.0
SQLAlchemy>=1.4.0
aiosqlite>=0.18.0
python-dotenv>=0.21.0
uvicorn>=0.17.0
aiohttp>=3.8.1
pydantic>=1.10.0
fastapi>=0.95.0
asyncpg>=0.27.0
